Timer fixes (#293)
* initial

* Small changes (#283)

* Small changes

* Update server.cpp

* remove dbg_timershift

* Add `timer::get_i64`

* Fix build error

* Fix build error II

* Small changes

* Small changes

* Optimize

* Add `convert_pitch_delta_to_non_linear_space`

* Update mouse.cpp

* Update mouse.cpp

* Small changes

* Change formatting

* Add `get_build_date`

* Fix a build error

* Small changes

* Use `GetTickCount64`

* Update destruction.cpp

* Update version.h

* Revert "Update version.h"

This reverts commit 8029caa0577c689d8fd14c9a823cceb25646fb8a.

* Update frametime.cpp

* Small changes

* Small fixes

* Add guard against backward QPC values

Added guard to prevent backward QPC values during VM live-migration.

* Improve `timer::get_i64`

* Update timer.cpp

* Update timer.cpp

* Small change

---------

+++ game_patch/multi/gametype.h
@@ -115,7 +115,7 @@ struct HillInfo
 
     // clientside visual smoothing
     bool vis_contested = false;
-    int vis_last_flip_ms = 0;
+    int64_t vis_last_flip_ms = 0;
 };
 
 struct CPGTRules // capture point game types
